feat: create squad modal
